+++ src/angularjs/src/app/analysis-jobs/create/analysis-jobs-create.controller.js
@@ -14,7 +14,7 @@
         var ctl = this;
 
         function initialize() {
-            ctl.neighborhoods = Neighborhood.all().$promise.then(function(data) {
+            ctl.neighborhoods = Neighborhood.all({ordering: 'label'}).$promise.then(function(data) {
                 ctl.neighborhoods = data.results;
             });
         }